Test/requirements.txt
--trusted-host ingress-platform.live-aws-useast1.bose.io
--extra-index-url https://ingress-platform.live-aws-useast1.bose.io/dev/svc-core-devops-nexus/prod/core-devops-nexus-core/repository/pypi-all/simple
CastleTestUtils>=0.1.7

pytest-testrail>=1.1.3
recordclass==0.5
